+++ src/components/calender/index.tsx
@@ -1,0 +1,3 @@
+import calender from './calender';
+
+export default calender;
