+++ src/index.ts
@@ -1,4 +1,4 @@
 import Input from './components/input';
-import Calender from './components/calender';
+import Calender from './components/calender/calender';
 
 export { Input, Calender };
